# swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "Google Ads keywordPlanner API",
    "version": "1.0.0",
    "title": "Google Ads API"
  },
  "host": "127.0.0.1:8080",
  "tags": [
    {
      "name": "keywordPlan",
      "description": "Everything about your keywordPlanner"
    }
  ],
  "schemes": [
    "http"
  ],
  "paths": {
    "/keywordPlan": {
      "get": {
        "tags": [
          "keywordPlan"
        ],
        "summary": "Fetch keywordPlan List",
        "description": "Fetch keywordPlan List",
        "operationId": "getKeywordPlanList",
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "integer",
                  "description": "The user ID."
                },
                "name": {
                  "type": "string",
                  "description": "name."
                },
                "resource_name": {
                  "type": "string",
                  "description": "resource name."
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "keywordPlan"
        ],
        "summary": "Create a new keywordPlan ",
        "description": "",
        "operationId": "createKeywordPlan",
        "responses": {
          "200": {
            "description": "Created"
          }
        }
      }
    },
    "/keywordPlan/{keywordPlanId}": {
      "get": {
        "tags": [
          "keywordPlan"
        ],
        "summary": "Find keywordPlan by ID",
        "description": "Returns a single keywordPlan",
        "operationId": "getKeywordPlan",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "keywordPlanId",
            "in": "path",
            "description": "ID of keywordPlan to return",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation"
          },
          "404": {
            "description": "keywordPlan not found"
          }
        }
      },
      "delete": {
        "tags": [
          "keywordPlan"
        ],
        "summary": "Deletes a keywordPlan",
        "description": "",
        "operationId": "deleteKeywordPlan",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "keywordPlanId",
            "in": "path",
            "description": "Pet id to delete",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid ID supplied"
          },
          "404": {
            "description": "Pet not found"
          }
        }
      }
    },
    "/keywordPlan/{keywordPlanId}/keywords": {
      "post": {
        "tags": [
          "keywordPlan"
        ],
        "summary": "Add keywords",
        "description": "Add keywords to keywordPlan",
        "operationId": "KeywordPlanKeywords",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "keywordPlanId",
            "in": "path",
            "description": "ID of keywordPlan to return",
            "required": true,
            "type": "integer",
            "format": "int64"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "array",
              "items": {
                "type": "string",
                "example": "купить машину"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation"
          },
          "404": {
            "description": "keywordPlan not found"
          }
        }
      }
    },
    "/keywordPlan/{keywordPlanId}/historicalMetrics": {
      "get": {
        "tags": [
          "keywordPlan"
        ],
        "summary": "Get Historical Metrics",
        "description": "Fetch historical metrics for keywordPlan",
        "operationId": "KeywordPlanMetrics",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "keywordPlanId",
            "in": "path",
            "description": "ID of keywordPlan to return",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation"
          },
          "404": {
            "description": "keywordPlan not found"
          }
        }
      }
    }
  },
  "externalDocs": {
    "description": "Find out more about Google Ads API",
    "url": "https://developers.google.com/google-ads/api/reference"
  }
}